news_notifier/requirements.txt
requests
newsapi-python
boto3
python-dotenv